Kafka/kafka-word-streaming/src/requirements.txt
kafka-python==2.0.2
Faker==14.1.0